fix(templates/drizzle-tidb): Move Drizzle database setup to index file
